Add files via upload
We used the RandMBin SAS macro to generate nine multivariate binary variables with specified means and an exchangeable correlation structure. We used these binary variables in a linear prediction model on the logit scale to produce binary exposure and outcome variables. Those binary variables served as the confounding variables as they explained the relationship between the exposure and outcome. We assumed that 50% of the cohort to be exposed and 10% to experience the outcome of interest. The macro was tested under the condition where the unadjusted association between the exposure and outcome variable was set at an odds ratio (OR) of 2.00, and the true (i.e., fully adjusted) OR was assumed to be 1.00. Finally, we misclassified each confounder by varying the sensitivity and specificity.