Add files via upload
The SAS macro initially checks for errors and ensures the required parameters are specified. Some parameters have default values, so there is no need for them to be specified for the macro to execute; however, we recommend the user review all default parameters to ensure appropriateness for their study. The program assumes exposure, outcome, and misclassified confounders are binary indicator variables coded as 0 or 1; an observation with the indicator “1” is considered a case and “0” is a non-case. All other non-misclassified independent variables entered in the model can be either binary or continuous variables. The logistic regression model in the macro can be modified to accommodate categorical variables with more than two levels; note, the macro is unable to correct for misclassified categorical variables.